Stub creating for the Quiz.java class
Quiz class has attributes for a title, a list of questions, and an array of answers. It has a constructor that takes a title and a list of questions as parameters, and initializes the answers array to have the same size as the list of questions. It also has getters and setters for each attribute, as well as a setAnswer() method to set the answer for a specific question.